test: adjust LuaJIT test suite for Tarantool
This patch makes it possible to run LuaJIT test suite using
Tarantool.

It disables these tests from the LuaJIT test suite,
because _G and some modules in Tarantool are different from _G and
modules in LuaJIT:
* 359 lib/contents.lua --- base
* 364 lib/contents.lua --- string
* 366 lib/contents.lua --- pre-5.2 table
* 370 lib/contents.lua --- os
* 371 lib/contents.lua --- debug
* 372 lib/contents.lua --- package
* 374 lib/contents.lua --- package.loaders
* 375 lib/contents.lua --- package.loaded

There is also some difference in Tarantool runtime (such as forbidden
global variables usage by default), so testing environment need to be
tweaked before the testing begins. For this purpose LUAJIT_TEST_INIT
CMake variable is introduced. This variable accepts the name of the Lua
script to be run prior to every test chunk.

Resolves tarantool/tarantool#4064
Part of tarantool/tarantool#4473

Co-authored-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Sergey Ostanevich <sergos@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>