test: adjust lua-Harness tests that use dofile
This patch makes out-of-source execution for lua-Harness suite tests
using `dofile()` for specific checks in corresponding *.t files correct.

The tests that use `dofile()` function on file in test sources directory
are listed below:
* 101-boolean.t
* 102-function.t
* 103-nil.t
* 104-number.t
* 105-string.t
* 106-table.t
* 107-thread.t
* 108-userdata.t
* 203-lexico.t
* 231-metatable.t
* 301-basic.t
* 305-utf8.t
* 404-ext.t

`dofile()` searches the files to execute in the current working
directory, that might be not the same as the test source directory.

This patch introduces the new function `make_specific_checks()` that
evaluates the path to file considering arg[0] (i.e. the given test
filename) value making it relative to the running test rather than to
the suite directory.

Part of tarantool/tarantool#5844
Part of tarantool/tarantool#4473

Co-authored-by: Igor Munkin <imun@tarantool.org>
Reviewed-by: Sergey Ostanevich <sergos@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>