FFI: Handle zero-fill of struct-of-NYI.
(cherry picked from ad65934fa08a65bfb0eb9528731a4394842cc173)

Trace recording is aborted with NYI reason for the case of a struct with
aggregate allocation. The general solution is way too complex due to
heterogeneous initialization of the structure. However, zero-fill case
requires nothing but "memset"-like initialization of GCcdata payload.

As a result of this patch workaround for the zero-fill structures is
introduced.

Sergey Kaplun:
* added the description and the test for the problem

Part of tarantool/tarantool#6548
Related to tarantool/tarantool#4630
Related to tarantool/tarantool#5885

Reviewed-by: Sergey Ostanevich <sergos@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>