test: disable PUC-Rio tests confused by -v output
Version and status are printed to stdout instead of stderr since
LuaJIT-2.0.0-beta11 (as it is not an error message).
See commit 0bd1a66f2f055211ef55834ccebca3b82d03c735 ('Print version and
JIT status to stdout, not stderr.').
This behavior is the same as in Lua 5.2.

This patch disables tests confused by unexpected -v output to stdout.

Relates to tarantool/tarantool#5687
Part of tarantool/tarantool#5845
Part of tarantool/tarantool#4473

Reviewed-by: Sergey Ostanevich <sergos@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>