Throw any errors before stack changes in trace stitching.
Thanks to doujiang24.

(cherry-picked from commit 3f9389edc6cdf3f78a6896d550c236860aed62b2)

The Lua stack is changed in the `recff_stitch`, so if the trace
is aborted, for example, when the `maxsnap` is reached, and an
error is thrown from there, the execution continues with an
unbalanced stack.

This patch adds a check for snapshot overflow to the
`recff_stitch` to overcome the issue for the `LJ_TRERR_SNAPOV`
case, but other cases remain unresolved and should be fixed in
other series. For a detailed description of the remaining cases,
see the issue #1166.

Maxim Kokryashkin:
* added the description and the test for the problem

Part of tarantool/tarantool#9595

Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Signed-off-by: Sergey Kaplun <skaplun@tarantool.org>