test: port lua-Harness to Test.Assertion
Backports conversion to Test.Assertion patch[1] from the
lua-Harness suite.

As a result of this change out-of-source testing is broken, but the
previous assertion using package.path as a second argument worked fine.
Because of that, this commit includes fix, that was made by this
patch[2] in the mainline lua-Harness.

[1]: https://framagit.org/fperrad/lua-Harness/-/commit/074fefa
[2]: https://framagit.org/fperrad/lua-Harness/-/commit/c4451fe

Part of tarantool/tarantool#5970
Part of tarantool/tarantool#4473

Reviewed-by: Igor Munkin <imun@tarantool.org>
Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>