ARM64: Reorder interpreter stack frame and fix unwinding.
Reported by Yichun Zhang. Fixes #722.
May help towards fixing #698, too.

(cherry picked from commit 421c4c798791d27b7f967df39891c4e4fa1d107c)

The `_Unwind_Find_FDE` fails to find the FDE (frame descriptor
element) for `lj_vm_ffi_call` in DWARF unwind info, despite
the presence of its data in the `.debug_frame` section.

LuaJIT emits its own DWARF entries for the CFI (call frame
information, section 6.4.1 in DWARF standard)[1].The FP
register value is vital to perform unwinding, and it is
possible to restore that register using the Canonical
Frame Address, or CFA. It can be obtained as `CFA - offset`.
By default, the CFA register is SP, however, it can be
changed to any other.

According to ARM's calling convention, the first eight
arguments of a function must be passed in x0-x7 registers,
and all the remaining must be passed on the stack. The
latter fact is important because it affects the SP and,
because of that, the CFA invalidates. This patch changes
the CFA register to the FP for the lj_vm_ffi_call, which
fixes the issue.

All the other changes are made just for refactoring purposes.

[1]: https://dwarfstd.org/doc/DWARF5.pdf

Maxim Kokryashkin:
* added the description and the test case for the problem

Needed for tarantool/tarantool#6096
Part of tarantool/tarantool#7230

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Sergey Ostanevich <sergos@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>