OSX/ARM64: Disable unwind info.
See #698.

(cherry picked from commit 78350a2565e1cf1102bcd25be406f02953d4dd3b)

External unwinding support is already disabled for OSX on ARM64
platform so there is no point in generation of incorrect unwind
info for it. This patch disables that generation for the commit
history match, and it will be re-enabled in the next commit,
which contains the fix for the issue.

Maxim Kokryashkin:
* added the description for the problem

Needed for tarantool/tarantool#6096
Part of tarantool/tarantool#7230

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Sergey Ostanevich <sergos@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>