test: separate memprof Lua API tests into subtests
As the number of memprof test cases is expected to grow,
memprof tests are separated into subtests to encapsulate
test cases and be able to skip some of the subtests.

Also, output generation and parsing are separated into a
dedicated function `generate_parsed_output`, which allows
one to run memprof on different payloads, passing a payload
funtion as an argument.

Part of tarantool/tarantool#5814

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>