memprof: group allocations on traces by traceno
When LuaJIT executes a trace, the trace number is stored as
the virtual machine state. So, we can treat this number as
an allocation event source in memprof and report allocation events
from traces as well.

Previously, all the allocations from traces were marked as INTERNAL.

This patch introduces the functionality described above by adding
a new allocation source type named ASOURCE_TRACE. If allocation event
occurs when trace is executed, trace number and trace's mcode starting
address are streamed to a binary file:

| loc-trace  := trace-no trace-addr
| trace-no   := <ULEB128>
| trace-addr := <ULEB128>

Also, the memory profiler parser is adjusted to recognize entries
mentioned above. The <loc> structure is extended with field <traceno>,
representing trace number. Trace locations are demangled as

| TRACE [<trace-no>] <trace-addr>

Resolves tarantool/tarantool#5814

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>