ARM64: Fix pcall() error case.
Reported by Alex Orlenko.

(cherry picked from commit b4b2dce9fc3ffaaaede39b36d06415311e2aa516)

The `pcall()` assembler preambule modifies `RC`, which is mapped to
`x28` (the value is N_args * 8), during the check of the amount of the
given arguments. So, this wrong value being used in the `fff_fallback`
routine leads to a crash on error throwing, because the Lua stack is
incorrectly filled and can't be unwound.

This patch adds the additional comparison before taking the fallback
branch and modifies `RC` only after this branch.

Sergey Kaplun:
* added the description and the test for the problem

Part of tarantool/tarantool#8069

Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>