test: adapt PUC-Rio tests for debug line hook
The LuaJIT's virtual machine interprets the bytecode following the
return from function (i.e. the one succeeding the call made) and located
on the line other than that return bytecode, as a new line trigger for
line hooks, unlike Lua does. Here is an example (it is joined in one
line intend):

| debug.sethook(function(_, l) print("LINE: "..l) end, "l") loadstring("\n\ns=nil")() debug.sethook()

This chunk prints for LuaJIT:

| LINE: 3
| LINE: 1

But for Lua 5.1 it is only "LINE: 3" in the output.

There are no answers in the corresponding thread[1] in lua-l@ lists, so
nobody can say for sure whether such difference is either Lua or LuaJIT
bug. Anyway, the affected tests are adapted to LuaJIT behavior by
changing assertions considering line events that LuaJIT generates.

[1]: http://lua-users.org/lists/lua-l/2022-03/msg00125.html

Resolves tarantool/tarantool#5693
Part of tarantool/tarantool#5870

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>