LJ_GC64: Fix HREFK optimization.
Contributed by XmiliaH.

(cherry-picked from commit 91bc6b8ad1f373c1ce9003dc024b2e21fad0e444)

In `lj_record_idx` when `ix->oldv` is the global nilnode and the
required key is not present in the table, it is possible to pass
the constant key lookup optimization condition because of the
`uint32_t` (`MSize`) overflow. Because of that, further recording
incorrectly removes the check for the nilnode, which produces
wrong results when trace is called for a different table.

The issue is solved by using `GCSize`, which has a size of
64 bits on GC64 builds, instead of `MSize`.

Maxim Kokryashkin:
* added the description and the test for the problem

Part of tarantool/tarantool#9145

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>