build: configure parallel jobs
lua-Harness and tarantool test suites uses prove(1) for running tests.
prove(1) allows to run tests in parallel with the option "--jobs" [1]
and this options requires argument, so we can't rely on default value.

It is not possible to get a number of parallel jobs in CMake passed by
the user with option "--parallel", but CMake allows to pass a number of
parallel jobs with environment variable CMAKE_BUILD_PARALLEL_LEVEL [2]
on configuration phase. The value set by that environment variable is
used as a number of CPU threads when it was not specified by user.
Number of CPU threads detected using builtin CMake function [3].

NOTE: CMAKE_BUILD_PARALLEL_LEVEL has been added in a version 3.12.
In earlier CMake versions environment variable is ignored.

1. https://perldoc.perl.org/prove
2. https://cmake.org/cmake/help/latest/envvar/CMAKE_BUILD_PARALLEL_LEVEL.html
3. https://cmake.org/cmake/help/latest/module/ProcessorCount.html

Co-authored-by: Igor Munkin <imun@tarantool.org>
Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>