test: increase sampling interval in sysprof tests
There are some issues faced in our CI when sysprof tests are run next to
another CPU bound workload. As a result a huge profile dump is generated
that is kinda QoD for CI runner. As a temporary solution it was decided
to relax profiler pressure by increasing profiling interval (the new
default is chosen as 0.1 seconds).

Needed for tarantool/tarantool#7472

Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>