test: make sysprof tests more deterministic
As a result of increasing the profiling interval in sysprof tests, the
assertion for LFUNC samples amount has become flaky, since there is no
guarantee a sample of LFUNC type will be found within such a sparse
profile. To make this test more deteministic, the assertion is rewritten
in a more rigid way: no JIT-related samples are collected, since JIT is
disabled; no non-Lua function (i.e. FFUNC and CFUNC) samples are
collected too; a sample of one of the other types has to be found in the
profiler dump.

Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>