test: introduce utils.profilename helper
Before the patch both memprof and sysprof artefacts were generated
within the binary artefacts tree (i.e. in the same directory LuaJIT
binary is generated). However, more convenient way is producing these
temporary profiles in a separate directory (e.g. located on the
partition with more strict space limits).

As a result of the patch <utils.profilename> helper is introduced and
all memprof and sysprof test chunks are patched to use it. The new
helper takes the basename of the profile to be created and also
considers LUAJIT_TEST_VARDIR environment variable to set the directory,
where test profiles are generated. For the case when LUAJIT_TEST_VARDIR
is not set, everything works as before.

Part of tarantool/tarantool#7472

Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>