build: introduce LUAJIT_ENABLE_CHECKHOOK option
Originally there is nether a special option nor a variable to configure
the check for instrunction/line hooks for compiled code via the build
system being used. We finally decided to use this feature in Tarantool,
so for convenient managing LUAJIT_ENABLE_CHECKHOOK option is added to
the root project CMakeLists.txt.

Needed for tarantool/tarantool#7762

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>